# Conflicts:
#	SuFood專案MVC/SuFood/Areas/BackStage/Views/BackHome/Coupon.cshtml
#	SuFood專案MVC/SuFood/Controllers/CheckOutController.cs
#	SuFood專案MVC/SuFood/SuFood.csproj
#	SuFood專案MVC/SuFood/Views/CheckOut/CheckOut.cshtml

+++ SuFood專案MVC/SuFood/Controllers/RetailController.cs
@@ -135,7 +135,13 @@ namespace SuFood.Controllers
 
             return Json(new { GetAccountId = getAccountId });
         }
+        [HttpGet]
+        public async Task<IActionResult> IsLogin()
+        {
+            var getAccountId = HttpContext.Session.GetString("GetAccountId");
 
+            return Json(new { GetAccountId = getAccountId });
+        }
         [HttpGet]
         public async Task<IEnumerable<ShoppingCartViewModel>> GetShoppingCarts()
         {
@@ -160,8 +166,8 @@ namespace SuFood.Controllers
 
             return viewModelList;
         }
-        [HttpPut]
-        public async Task<IEnumerable<ShoppingCartViewModel>> AddCartItem([FromBody] ShoppingCartViewModel model)
+        [HttpPost]
+        public async Task<bool> AddCartItem([FromBody] ShoppingCartViewModel model)
         {
             var getAccountId = HttpContext.Session.GetString("GetAccountId");
 
@@ -170,26 +176,28 @@ namespace SuFood.Controllers
 
             if (cartItem == null)
             {
-                return null;
+                return false;
             }
 
             cartItem.CartQuantity++;
             _context.ShoppingCart.Update(cartItem);
             await _context.SaveChangesAsync();
-            await GetShoppingCarts();
-            return (IEnumerable<ShoppingCartViewModel>)cartItem;
+            return true;
         }
 
 
-        [HttpPut]
-        public async Task<IEnumerable<ShoppingCartViewModel>> MinusCartItem([FromBody] ShoppingCartViewModel model)
+        [HttpPost]
+        public async Task<bool> MinusCartItem([FromBody] ShoppingCartViewModel model)
         {
             var getAccountId = HttpContext.Session.GetString("GetAccountId");
 
             var cartItem = await _context.ShoppingCart
                 .FirstOrDefaultAsync(c => c.ProductId == model.ProductId && c.AccountId == Convert.ToInt32(getAccountId));
 
-
+            if (cartItem == null)
+            {
+                return false;
+            }
             cartItem.CartQuantity--;
 
             if (cartItem.CartQuantity == 0)
@@ -203,23 +211,58 @@ namespace SuFood.Controllers
 
             await _context.SaveChangesAsync();
             await GetShoppingCarts();
-            return (IEnumerable<ShoppingCartViewModel>)cartItem;
+            return true;
         }
-        [HttpPut]
-        public async Task<IEnumerable<ShoppingCartViewModel>> DeleteCartItem([FromBody] ShoppingCartViewModel model)
+        [HttpPost]
+        public async Task<bool> DeleteCartItem([FromBody] ShoppingCartViewModel model)
         {
             var getAccountId = HttpContext.Session.GetString("GetAccountId");
 
             var cartItem = await _context.ShoppingCart
                 .FirstOrDefaultAsync(c => c.ProductId == model.ProductId && c.AccountId == Convert.ToInt32(getAccountId));
-
+            if (cartItem == null)
+            {
+                return false;
+            }
 
             _context.ShoppingCart.Remove(cartItem);
 
             await _context.SaveChangesAsync();
             await GetShoppingCarts();
-            return (IEnumerable<ShoppingCartViewModel>)cartItem;
+            return true;
         }
+        [HttpPost]
+        public async Task<IActionResult> AddSingle2Cart([FromBody] CartViewModels model)
+        {
+            var getAccountId = HttpContext.Session.GetString("GetAccountId");
 
+            if (getAccountId == null)
+            {
+                return Json(new { GetAccountId = "null" });
+            }
+
+            var existingCart = await _context.ShoppingCart
+        .FirstOrDefaultAsync(c => c.ProductId == model.ProductId && c.AccountId == Convert.ToInt32(getAccountId));
+
+            if (existingCart != null)
+            {
+                existingCart.CartQuantity += model.CartQuantity;
+                await _context.SaveChangesAsync();
+            }
+            else
+            {
+                ShoppingCart cart = new ShoppingCart
+                {
+                    CartId = model.CartId,
+                    AccountId = Convert.ToInt32(getAccountId),
+                    CartQuantity = model.CartQuantity,
+                    ProductId = model.ProductId,
+                };
+                _context.ShoppingCart.Add(cart);
+                await _context.SaveChangesAsync();
+            }
+
+            return Json(new { GetAccountId = getAccountId });
+        }
     }
 }
